// src/app/models/all-queries-data-row.ts
export class AllQueriesDataRow {
    id: number;
    query: string;
    click: number;
    impression: number;
    inputed_ctr: number;
    position: number;
    ctr: number;
    instance: number;
    expected_ctr: number;
    ctr_delta: number;
    expected_clicks: number;
    traffic_loss: number;
    traffic_gain: number;
    nr1: number;
    nr2: number;
    nr3: number;
    nr4: number;
    nr5: number;
    nr6: number;
    nr7: number;
    nr8: number;
    nr9: number;
    nr10: number;
    nr11: number;
}